v.0.2: Added colorama, changed to millisec, added example_usage
Code:
Added a try import for colorama.
Added color to timmyprint
Added timmyp as alias for timmyprint
Now prints ms and not sec.
Changed print string/format.

Setup:
Updated version so this can be updated on pypi.
Added readme as longdescription
Added classifiers

Tests:
Not exactly a test, but added an example usage file.

Readme:
Added example_usage as an example in the readme.